Week 4, Leson 6, Assignment 4, Task 1 - Peer Graded: Client-Server Communication
Assignment Overview

In this assignment, you will update the web application to get data from the server to render the information corresponding to the promotions and the leadership team of the company. In addition, you will handle the submission of the feedback form by posting the feedback to the server. At the end of this assignment, you should have completed the following:

- Updated the promotion item in the home page to be rendered based on the data downloaded from the server.
- Updated the home page and the aboutus page to render the information about the corporate leaders by downloading the data from the server
- Enabled the users to submit feedback through the feedback form and recording their feedback on the server.

Assignment Requirements

Task 1

In this task, you will update the code in services.js and controllers.js and home.html:

- Update getPromotion() to return a resource that enables the fetching of the promotion data from the server
- Update the IndexController to render the promotion data. You should handle the error condition appropriately.
- Update home.html to render the promotion data when it is correctlyreceived, and error message otherwise.

 Changes to be committed:
	modified:   app/scripts/controllers.js
	modified:   app/scripts/services.js
	modified:   app/views/home.html