fix bug 'gulp watch' --> + port: 2999 in gulpfile.js
Week 4, Lesson 6, Assignment 4, Task 2 - Peer Graded Assignment: Client-Server Communication

Assignment Overview

In this assignment, you will update the web application to get data from the server to render the information corresponding to the promotions and the leadership team of the company. In addition, you will handle the submission of the feedback form by posting the feedback to the server. At the end of this assignment, you should have completed the following:

- Updated the promotion item in the home page to be rendered based on the data downloaded from the server.
- Updated the home page and the aboutus page to render the information about the corporate leaders by downloading the data from the server
- Enabled the users to submit feedback through the feedback form and recording their feedback on the server.

Assignment Requirements

Task 2

In this task, you will update the code in services.js, controllers.js and home.html and aboutus.html. You will handle the fetching and rendering of the leadership data in both home page and aboutus page:

- Update corporateFactory to return the resource for the leadership
- Update IndexController and AboutController to render the leadership data
  obtained from the server, and error condition should be handled.
- Update home.html and aboutus.html to render the leadership data when it is correctly received, and error message otherwise.

Changes to be committed:
	modified:   gulpfile.js